requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy
pandas
Pillow
nltk
matplotlib